+++ database/lang_pl.php
@@ -224,4 +224,5 @@ $lang['live_stat_subs'] = "Zmiany";
 $lang['live_replay'] = "Powtórka";
 $lang['live_replay_button'] = "Powtórka na telebimie";
 $lang['live_replay_sent'] = "Powtórka wysłana na telebim.";
+$lang['live_search_player'] = "Szukaj zawodnika…";
 ?>